+++ packages/codered-agent/src/errors/AgentUnpairedError.ts
@@ -1,0 +1,6 @@
+export class AgentUnpairedError extends Error {
+  public constructor(message = 'Agent is unpaired') {
+    super(message);
+    this.name = 'AgentUnpairedError';
+  }
+}